Create IconButton and LoginButton variations of Button
