+++ src/components/Button/Button.css
@@ -38,17 +38,3 @@
   border: 3px solid transparent;
   color: #c0c0c0;
 }
-
-.iconLeft {
-  width: 2rem;
-  margin-right: 0.5rem;
-}
-
-.iconRight {
-  width: 2rem;
-  margin-left: 0.5rem;
-}
-
-a .button:hover {
-   
-}
