+++ src/components/Button/IconButton.css
@@ -1,0 +1,9 @@
+.iconLeft {
+  width: 2rem;
+  margin-right: 0.5rem;
+}
+
+.iconRight {
+  width: 2rem;
+  margin-left: 0.5rem;
+}
